Refactor component exports
