+++ package.json
@@ -23,6 +23,7 @@
     "@fontsource/cabin": "^4.5.0",
     "@types/jsonp": "^0.2.1",
     "@types/react": "^18.0.9",
+    "@types/react-helmet": "^6.1.5",
     "@types/styled-components": "^5.1.25",
     "babel-plugin-styled-components": "^1.13.2",
     "date-fns": "^2.23.0",
