+++ src/components/index.tsx
@@ -1,0 +1,17 @@
+export { Button } from "./Button";
+export { SmallButton } from "./Button";
+export { Card } from "./Card";
+export { CardDeck } from "./CardDeck";
+export { committee } from "./committee";
+export { CommitteeCard } from "./CommitteeCard";
+export { EventsDeck } from "./EventsDeck";
+export { FadeIn } from "./FadeIn";
+export { Footer } from "./Footer";
+export { Layout } from "./Layout";
+export { Navbar } from "./Navbar";
+export { NavItem } from "./NavItem";
+export { Particles } from "./Particles";
+export { PostLink } from "./PostLink";
+export { RegistrationForm } from "./RegistrationForm";
+export { TagLink } from "./Tag";
+export { TextLink } from "./TextLink";
