+++ src/pages/about.tsx
@@ -3,8 +3,7 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 
 // Custom Components
-import Layout from "~/components/Layout";
-import CommitteeCard from "~/components/CommitteeCard";
+import { Layout, CommitteeCard } from "~/components";
 
 // Import committee information to be converted into displayable components.
 import { committee } from "~/content/committee/history";
