+++ src/pages/blog.tsx
@@ -3,9 +3,7 @@ import { graphql } from "gatsby";
 import PropTypes from "prop-types";
 
 // Components
-import Layout from "~/components/Layout";
-import PostLink from "~/components/PostLink";
-import TextLink from "~/components/TextLink";
+import { Layout, PostLink, TextLink } from "~/components";
 
 const BlogIndexPage = ({
   data: {
