+++ src/pages/contact.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 
-import Button from "~/components/Button";
-import Layout from "~/components/Layout";
+import { Button, Layout } from "~/components";
 
 const Contact = () => (
   <Layout title="Contact Us">
