+++ src/pages/newsletter.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import Layout from "~/components/Layout";
-import { RegistrationForm } from "~/components/RegistrationForm";
+import { Layout, RegistrationForm } from "~/components";
 
 import { color } from "~/theme/config";
